perf(web): load the OpenTelemetry SDK on demand
`telemetry.ts` statically imported eight `@opentelemetry/*` packages, so the
whole tracing SDK shipped in the eagerly-loaded entry chunk of every build —
including `@opentelemetry/context-zone`, which pulls in zone.js. zone.js
monkeypatches the global async primitives (promises, timers, event listeners)
as soon as it is evaluated, so the cost was not only bytes but startup work on
every page load.

Tracing is opt-in on `VITE_OTEL_EXPORTER_OTLP_ENDPOINT`, which is unset in
ordinary production builds, so that work was pure waste there. The module's
own comment claimed builds without a collector were "unaffected" — true at
runtime, never true for the bundle.

Move the provider wiring to `telemetryProvider.ts` and reach it through a
dynamic `import()` behind the existing endpoint check. The check stays
synchronous, so an unconfigured build resolves without fetching or evaluating
any of it.

`initBrowserTelemetry()` now returns a promise. `main.tsx` chains the identity
probe off it so that first request is still traced when tracing is on;
`resolveIdentity` memoizes its promise, so the kick is only a warm-up and
later callers await the same request.
